Migrate NetworkCheckerToolPanel business code
